requirements.txt
fastmcp>=0.1.0
fastapi>=0.100.0
pyodbc>=5.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0